+++ client/flutter/lib/models/question.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/foundation.dart';
+
+class Question {
+  String title;
+  String body;
+
+  Question({@required this.title, @required this.body});
+}
