+++ client/flutter/pubspec.yaml
@@ -25,8 +25,9 @@ dependencies:
   yaml: ^2.2.0
   http: ^0.12.0+3
   uuid: 2.0.4
-  package_info: '>=0.4.0+16 <2.0.0'
+  package_info: ">=0.4.0+16 <2.0.0"
   flutter_svg: ^0.17.3+1
+  provider: ^4.0.4
 
   intl: ^0.16.0
   flutter_localizations:
